requirements.txt
uvicorn
fastapi
pypdf
numpy
python-multipart
langchain-openai
langchain_groq 
langchain_chroma
langchain-community
langchain_huggingface
langchain_openai
langchain_groq
scikit-learn